+++ packages/blitz/src/cli/utils/next-commands.ts
@@ -1,4 +1,3 @@
-import {readBlitzConfig} from "../../index-server"
 import {normalize, ServerConfig} from "./config"
 import {
   nextBuild,
@@ -8,12 +7,24 @@ import {
   startCustomServer,
   buildCustomServer,
 } from "./next-utils"
+import * as fs from "fs"
+import * as path from "path"
 
 export async function build(config: ServerConfig) {
   const {rootFolder, nextBin, watch} = await normalize(config)
 
   await nextBuild(nextBin, rootFolder, {} as any, config)
   if (customServerExists()) await buildCustomServer({watch})
+}
+
+export function readBlitzConfig(rootFolder: string = process.cwd()) {
+  const packageJsonFile = fs.readFileSync(path.join(rootFolder, "package.json"), {
+    encoding: "utf8",
+    flag: "r",
+  })
+  const packageJson = JSON.parse(packageJsonFile)
+
+  return packageJson.blitz || {}
 }
 
 export async function dev(config: ServerConfig) {
